protect sensitive endpoints

+++ api/user.go
@@ -1,6 +1,7 @@
 package api
 
 import (
+	"errors"
 	"net/http"
 
 	"github.com/hamster2020/gauth"
@@ -30,6 +31,12 @@ func createUser(logic gauth.Logic) http.HandlerFunc {
 			return
 		}
 
+		caller := getCallerFromRequest(req)
+		if r.Roles.HasRole(gauth.RolesAdmin) && !caller.HasRole(gauth.RolesAdmin) {
+			writeJSONError(crw, errors.New("only admin can create a new admin user"), http.StatusBadRequest)
+			return
+		}
+
 		err := logic.CreateUser(r)
 		if err != nil {
 			writeJSONError(crw, err, http.StatusInternalServerError)
@@ -45,6 +52,12 @@ func userByEmail(logic gauth.Logic) http.HandlerFunc {
 		crw := w.(*codeResponseWriter)
 
 		email := req.URL.Query().Get(":email")
+		caller := getCallerFromRequest(req)
+		if caller.Info() != email && !caller.HasRole(gauth.RolesAdmin) {
+			writeJSONError(crw, errors.New("non-admin users can only look up their own user account"), http.StatusBadRequest)
+			return
+		}
+
 		user, err := logic.UserByEmail(email)
 		if err != nil {
 			writeJSONError(crw, err, http.StatusInternalServerError)
@@ -66,6 +79,17 @@ func updateUser(logic gauth.Logic) http.HandlerFunc {
 		}
 
 		email := req.URL.Query().Get(":email")
+		caller := getCallerFromRequest(req)
+		if caller.Info() != email && !caller.HasRole(gauth.RolesAdmin) {
+			writeJSONError(crw, errors.New("non-admin users can only update their own user account"), http.StatusBadRequest)
+			return
+		}
+
+		if r.Roles.HasRole(gauth.RolesAdmin) && !caller.HasRole(gauth.RolesAdmin) {
+			writeJSONError(crw, errors.New("non-admin users update their roles to include admin users"), http.StatusBadRequest)
+			return
+		}
+
 		if err := logic.UpdateUser(email, r); err != nil {
 			writeJSONError(crw, err, http.StatusInternalServerError)
 			return
@@ -80,6 +104,12 @@ func deleteUser(logic gauth.Logic) http.HandlerFunc {
 		crw := w.(*codeResponseWriter)
 
 		email := req.URL.Query().Get(":email")
+		caller := getCallerFromRequest(req)
+		if caller.Info() != email && !caller.HasRole(gauth.RolesAdmin) {
+			writeJSONError(crw, errors.New("non-admin users can only delete their own user account"), http.StatusBadRequest)
+			return
+		}
+
 		if err := logic.DeleteUser(email); err != nil {
 			writeJSONError(crw, err, http.StatusInternalServerError)
 			return
